Flickr-miner: Synchronous version with pooling
Since GdMiner object frees job after end of query_* function, the query
needs to do everything synchrounously. There were a problem with doing
it all synchrounously because of asynchronous adding of sources. This is
not still solved. This version uses GThreadPool for browsing containers.
It hangs when it comes to browsing not-root containers, which it did not
do when it was asynchronous or when the functions were used standalone.